Split administration client tests into multiple files
