+++ sdk/formrecognizer/Azure.AI.FormRecognizer/tests/DocumentModelAdministrationClient/DocumentModelAdministrationLiveTests.cs
@@ -11,29 +11,31 @@ using NUnit.Framework;
 namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
 {
     /// <summary>
-    /// The suite of tests for the <see cref="DocumentModelAdministrationClient"/> class.
+    /// The suite of tests for the document model methods in the <see cref="DocumentModelAdministrationClient"/> class.
     /// </summary>
     /// <remarks>
     /// These tests have a dependency on live Azure services and may incur costs for the associated
     /// Azure subscription.
     /// </remarks>
     [IgnoreServiceError(400, "InvalidRequest", Message = "Content is not accessible: Invalid data URL", Reason = "https://github.com/Azure/azure-sdk-for-net/issues/28923")]
-    public class DocumentModelAdministrationClientLiveTests : DocumentAnalysisLiveTestBase
+    public class DocumentModelAdministrationLiveTests : DocumentAnalysisLiveTestBase
     {
-        private readonly IReadOnlyDictionary<string, string> TestingTags = new Dictionary<string, string>()
+        private readonly IReadOnlyDictionary<string, string> _testingTags = new Dictionary<string, string>()
         {
             { "ordinary tag", "an ordinary tag" },
             { "crazy tag", "a CRAZY tag 123!@#$%^&*()[]{}\\/?.,<>" }
         };
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="DocumentModelAdministrationClientLiveTests"/> class.
+        /// Initializes a new instance of the <see cref="DocumentModelAdministrationLiveTests"/> class.
         /// </summary>
         /// <param name="isAsync">A flag used by the Azure Core Test Framework to differentiate between tests for asynchronous and synchronous methods.</param>
-        public DocumentModelAdministrationClientLiveTests(bool isAsync, DocumentAnalysisClientOptions.ServiceVersion serviceVersion)
+        public DocumentModelAdministrationLiveTests(bool isAsync, DocumentAnalysisClientOptions.ServiceVersion serviceVersion)
             : base(isAsync, serviceVersion)
         {
         }
+
+        #region Build
 
         [RecordedTest]
         public async Task BuildModelCanAuthenticateWithTokenCredential()
@@ -49,8 +51,6 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
 
             await client.DeleteDocumentModelAsync(modelId);
         }
-
-        #region BuildModel
 
         [RecordedTest]
         [TestCase(true)]
@@ -145,7 +145,7 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
 
             var options = new BuildDocumentModelOptions();
 
-            foreach (var tag in TestingTags)
+            foreach (var tag in _testingTags)
             {
                 options.Tags.Add(tag);
             }
@@ -154,7 +154,7 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
 
             DocumentModelDetails model = operation.Value;
 
-            CollectionAssert.AreEquivalent(TestingTags, model.Tags);
+            CollectionAssert.AreEquivalent(_testingTags, model.Tags);
         }
 
         [RecordedTest]
@@ -169,141 +169,10 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
             Assert.AreEqual("InvalidArgument", ex.ErrorCode);
         }
 
-        #endregion BuildModel
+        #endregion
 
-        #region management ops
+        #region Copy
 
-        [RecordedTest]
-        public async Task GetResourceDetails()
-        {
-            var client = CreateDocumentModelAdministrationClient();
-
-            ResourceDetails resourceDetails = await client.GetResourceDetailsAsync();
-
-            Assert.IsNotNull(resourceDetails.CustomDocumentModelCount);
-            Assert.IsNotNull(resourceDetails.CustomDocumentModelLimit);
-        }
-
-        [RecordedTest]
-        public async Task GetPrebuiltModel()
-        {
-            var client = CreateDocumentModelAdministrationClient();
-
-            DocumentModelDetails model = await client.GetDocumentModelAsync("prebuilt-businessCard");
-
-            ValidateDocumentModelDetails(model);
-            Assert.NotNull(model.Description);
-        }
-
-        [RecordedTest]
-        [TestCase(true)]
-        [TestCase(false)]
-        public async Task AdminOps(bool useTokenCredential)
-        {
-            var client = CreateDocumentModelAdministrationClient(useTokenCredential);
-            var trainingFilesUri = new Uri(TestEnvironment.BlobContainerSasUrl);
-            var modelId = Recording.GenerateId();
-            var description = "This is a test model.";
-
-            var options = new BuildDocumentModelOptions()
-            {
-                Description = description
-            };
-
-            foreach (var tag in TestingTags)
-            {
-                options.Tags.Add(tag);
-            }
-
-            BuildDocumentModelOperation operation = await client.BuildDocumentModelAsync(WaitUntil.Completed, trainingFilesUri, DocumentBuildMode.Template, modelId, options: options);
-            DocumentModelDetails resultModel = await client.GetDocumentModelAsync(modelId);
-
-            ValidateDocumentModelDetails(resultModel, description, TestingTags);
-
-            DocumentModelSummary modelSummary = client.GetDocumentModelsAsync().ToEnumerableAsync().Result
-                .FirstOrDefault(m => m.ModelId == modelId);
-
-            Assert.NotNull(modelSummary);
-
-            ValidateDocumentModelSummary(modelSummary, description, TestingTags);
-
-            await client.DeleteDocumentModelAsync(modelId);
-
-            Assert.ThrowsAsync<RequestFailedException>(() => client.GetDocumentModelAsync(modelId));
-        }
-
-        [RecordedTest]
-        public void DeleteModelFailsWhenModelDoesNotExist()
-        {
-            var client = CreateDocumentModelAdministrationClient();
-            var fakeModelId = "00000000-0000-0000-0000-000000000000";
-
-            RequestFailedException ex = Assert.ThrowsAsync<RequestFailedException>(async () => await client.DeleteDocumentModelAsync(fakeModelId));
-            Assert.AreEqual("NotFound", ex.ErrorCode);
-        }
-
-        [RecordedTest]
-        [TestCase(true)]
-        [TestCase(false)]
-        public async Task GetAndListOperations(bool useTokenCredential)
-        {
-            var client = CreateDocumentModelAdministrationClient(useTokenCredential);
-
-            // Guarantee there is going to be at least one operation
-            var modelId = Recording.GenerateId();
-            await using var trainedModel = await BuildDisposableDocumentModelAsync(modelId);
-
-            var modelOperationFromList = client.GetOperationsAsync().ToEnumerableAsync().Result;
-            Assert.GreaterOrEqual(modelOperationFromList.Count, 1);
-
-            ValidateOperationSummary(modelOperationFromList.FirstOrDefault());
-
-            OperationDetails operationDetails = await client.GetOperationAsync(modelOperationFromList.FirstOrDefault().OperationId);
-
-            ValidateOperationDetails(operationDetails);
-        }
-
-        [RecordedTest]
-        public async Task GetAndListOperationsWithTags()
-        {
-            var client = CreateDocumentModelAdministrationClient();
-            var trainingFilesUri = new Uri(TestEnvironment.BlobContainerSasUrl);
-            var modelId = Recording.GenerateId();
-            var options = new BuildDocumentModelOptions();
-
-            foreach (var tag in TestingTags)
-            {
-                options.Tags.Add(tag);
-            }
-
-            BuildDocumentModelOperation operation = await client.BuildDocumentModelAsync(WaitUntil.Started, trainingFilesUri, DocumentBuildMode.Template, modelId, options: options);
-
-            OperationSummary operationSummary = client.GetOperationsAsync().ToEnumerableAsync().Result
-                .FirstOrDefault(op => op.OperationId == operation.Id);
-
-            Assert.NotNull(operationSummary);
-
-            CollectionAssert.AreEquivalent(TestingTags, operationSummary.Tags);
-
-            OperationDetails operationDetails = await client.GetOperationAsync(operation.Id);
-
-            CollectionAssert.AreEquivalent(TestingTags, operationDetails.Tags);
-
-            await client.DeleteDocumentModelAsync(modelId);
-        }
-
-        [RecordedTest]
-        public void GetOperationWrongId()
-        {
-            var client = CreateDocumentModelAdministrationClient();
-
-            RequestFailedException ex = Assert.ThrowsAsync<RequestFailedException>(() => client.GetOperationAsync("0000000000000000"));
-            Assert.AreEqual("NotFound", ex.ErrorCode);
-        }
-
-        #endregion management ops
-
-        #region copy
         [RecordedTest]
         [TestCase(true)]
         [TestCase(false)]
@@ -337,21 +206,6 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
         }
 
         [RecordedTest]
-        public async Task GetCopyAuthorization()
-        {
-            var targetClient = CreateDocumentModelAdministrationClient();
-            var modelId = Recording.GenerateId();
-
-            DocumentModelCopyAuthorization targetAuth = await targetClient.GetCopyAuthorizationAsync(modelId);
-
-            Assert.IsNotNull(targetAuth.TargetModelId);
-            Assert.AreEqual(modelId, targetAuth.TargetModelId);
-            Assert.IsNotNull(targetAuth.AccessToken);
-            Assert.IsNotNull(targetAuth.TargetResourceId);
-            Assert.IsNotNull(targetAuth.TargetResourceRegion);
-        }
-
-        [RecordedTest]
         public async Task CopyModelToWithTags()
         {
             var sourceClient = CreateDocumentModelAdministrationClient();
@@ -360,7 +214,7 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
 
             await using var trainedModel = await BuildDisposableDocumentModelAsync(modelId);
 
-            var tags = TestingTags.ToDictionary(t => t.Key, t => t.Value);
+            var tags = _testingTags.ToDictionary(t => t.Key, t => t.Value);
 
             var targetModelId = Recording.GenerateId();
             DocumentModelCopyAuthorization targetAuth = await targetClient.GetCopyAuthorizationAsync(targetModelId, tags: tags);
@@ -368,7 +222,7 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
 
             DocumentModelDetails copiedModel = operation.Value;
 
-            CollectionAssert.AreEquivalent(TestingTags, copiedModel.Tags);
+            CollectionAssert.AreEquivalent(_testingTags, copiedModel.Tags);
 
             await sourceClient.DeleteDocumentModelAsync(targetModelId);
         }
@@ -386,9 +240,9 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
             Assert.AreEqual("InvalidRequest", ex.ErrorCode);
         }
 
-        #endregion copy
+        #endregion Copy
 
-        #region Compose model
+        #region Compose
 
         [RecordedTest]
         [TestCase(false)]
@@ -437,14 +291,14 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
             await using var trainedModelB = await BuildDisposableDocumentModelAsync(modelBId);
 
             var modelIds = new List<string> { trainedModelA.ModelId, trainedModelB.ModelId };
-            var tags = TestingTags.ToDictionary(t => t.Key, t => t.Value);
+            var tags = _testingTags.ToDictionary(t => t.Key, t => t.Value);
 
             var composedModelId = Recording.GenerateId();
             ComposeDocumentModelOperation operation = await client.ComposeDocumentModelAsync(WaitUntil.Completed, modelIds, composedModelId, tags: tags);
 
             DocumentModelDetails composedModel = operation.Value;
 
-            CollectionAssert.AreEquivalent(TestingTags, composedModel.Tags);
+            CollectionAssert.AreEquivalent(_testingTags, composedModel.Tags);
 
             await client.DeleteDocumentModelAsync(composedModelId);
         }
@@ -465,86 +319,53 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
             Assert.AreEqual("InvalidRequest", ex.ErrorCode);
         }
 
-        #endregion Composed model
+        #endregion
 
-        private void ValidateOperationSummary(OperationSummary operationSummary)
+        #region Get
+
+        [RecordedTest]
+        public async Task GetPrebuiltModel()
         {
-            Assert.NotNull(operationSummary.OperationId);
-            Assert.AreNotEqual(default(DateTimeOffset), operationSummary.CreatedOn);
-            Assert.AreNotEqual(default(DateTimeOffset), operationSummary.LastUpdatedOn);
-            Assert.AreNotEqual(default(DocumentOperationKind), operationSummary.Kind);
-            Assert.NotNull(operationSummary.ResourceLocation);
+            var client = CreateDocumentModelAdministrationClient();
 
-            if (operationSummary.Status == DocumentOperationStatus.Succeeded)
-            {
-                Assert.AreEqual(100, operationSummary.PercentCompleted);
-            }
+            DocumentModelDetails model = await client.GetDocumentModelAsync("prebuilt-businessCard");
+
+            ValidateDocumentModelDetails(model);
+            Assert.NotNull(model.Description);
         }
 
-        private void ValidateOperationDetails(OperationDetails operationDetails)
+        #endregion
+
+        #region Delete
+
+        [RecordedTest]
+        public void DeleteModelFailsWhenModelDoesNotExist()
         {
-            Assert.NotNull(operationDetails.OperationId);
-            Assert.AreNotEqual(default(DateTimeOffset), operationDetails.CreatedOn);
-            Assert.AreNotEqual(default(DateTimeOffset), operationDetails.LastUpdatedOn);
-            Assert.AreNotEqual(default(DocumentOperationKind), operationDetails.Kind);
-            Assert.NotNull(operationDetails.ResourceLocation);
+            var client = CreateDocumentModelAdministrationClient();
+            var fakeModelId = "00000000-0000-0000-0000-000000000000";
 
-            if (operationDetails.Status == DocumentOperationStatus.Succeeded)
-            {
-                Assert.AreEqual(100, operationDetails.PercentCompleted);
+            RequestFailedException ex = Assert.ThrowsAsync<RequestFailedException>(async () => await client.DeleteDocumentModelAsync(fakeModelId));
+            Assert.AreEqual("NotFound", ex.ErrorCode);
+        }
 
-                DocumentModelDetails result = operationDetails switch
-                {
-                    DocumentModelBuildOperationDetails buildOp => buildOp.Result,
-                    DocumentModelCopyToOperationDetails copyToOp => copyToOp.Result,
-                    DocumentModelComposeOperationDetails composeOp => composeOp.Result,
-                    _ => null
-                };
+        #endregion
 
-                if (result != null)
-                {
-                    ValidateDocumentModelDetails(result);
-                }
-            }
-            else if (operationDetails.Status == DocumentOperationStatus.Failed)
-            {
-                Assert.NotNull(operationDetails.Error);
-                Assert.NotNull(operationDetails.Error.Code);
-                Assert.NotNull(operationDetails.Error.Message);
-            }
+        [RecordedTest]
+        public async Task GetCopyAuthorization()
+        {
+            var targetClient = CreateDocumentModelAdministrationClient();
+            var modelId = Recording.GenerateId();
+
+            DocumentModelCopyAuthorization targetAuth = await targetClient.GetCopyAuthorizationAsync(modelId);
+
+            Assert.IsNotNull(targetAuth.TargetModelId);
+            Assert.AreEqual(modelId, targetAuth.TargetModelId);
+            Assert.IsNotNull(targetAuth.AccessToken);
+            Assert.IsNotNull(targetAuth.TargetResourceId);
+            Assert.IsNotNull(targetAuth.TargetResourceRegion);
         }
 
         private void ValidateDocumentModelDetails(DocumentModelDetails model, string description = null, IReadOnlyDictionary<string, string> tags = null)
-        {
-            if (description != null)
-            {
-                Assert.AreEqual(description, model.Description);
-            }
-
-            if (tags != null)
-            {
-                CollectionAssert.AreEquivalent(tags, model.Tags);
-            }
-
-            Assert.IsNotNull(model.ModelId);
-            Assert.AreNotEqual(default(DateTimeOffset), model.CreatedOn);
-
-            if (_serviceVersion >= DocumentAnalysisClientOptions.ServiceVersion.V2023_02_28_Preview)
-            {
-                if (model.ExpiresOn.HasValue)
-                {
-                    Assert.Greater(model.ExpiresOn, model.CreatedOn);
-                }
-            }
-            else
-            {
-                Assert.Null(model.ExpiresOn);
-            }
-
-            // TODO add validation for Doctypes https://github.com/Azure/azure-sdk-for-net-pr/issues/1432
-        }
-
-        private void ValidateDocumentModelSummary(DocumentModelSummary model, string description = null, IReadOnlyDictionary<string, string> tags = null)
         {
             if (description != null)
             {
