+++ sdk/formrecognizer/assets.json
@@ -2,5 +2,5 @@
   "AssetsRepo": "Azure/azure-sdk-assets",
   "AssetsRepoPrefixPath": "net",
   "TagPrefix": "net/formrecognizer",
-  "Tag": "net/formrecognizer_2ca9212f98"
+  "Tag": "net/formrecognizer_564a6da924"
 }
